+++ wattpad_scraper/utils/__init__.py
@@ -1,1 +1,2 @@
 from .request import get
+from .log import Log