+++ CHANGELOG.md
@@ -88,3 +88,16 @@ Auth System (Beta)
 - Get book by url
 - Convert books to epub
 - Login to wattpad
+
+## [0.0.25]
+* Fixed verbose mode. Added system to catch responses so it run faster on sencond run.
+### Major Changes:
+- Catch responses now run faster.
+### Minor Changes:
+- Added verbose mode.
+
+### Features
+- Search books
+- Get book by url
+- Convert books to epub
+- Login to wattpad